+++ src/FolkerKinzel.Strings/md/ReleaseNotes.md
@@ -1,4 +1,5 @@
-﻿- Polyfills for the static System.Buffers.SearchValues class and the System.Buffers.SearchValues<char> class (.NET 7.0, .NET 6.0, .NET 5.0, .NET Core 3.1, .NET Standard 2.1, .NET Standard 2.0, .NET Framework 4.6.1):
+﻿- The new `PersistentStringHash` struct allows to ombine the hash code for multiple character based values into a single persistent hash code.
+- Polyfills for the static System.Buffers.SearchValues class and the System.Buffers.SearchValues<char> class (.NET 7.0, .NET 6.0, .NET 5.0, .NET Core 3.1, .NET Standard 2.1, .NET Standard 2.0, .NET Framework 4.6.1):
 ```csharp
 static class FolkerKinzel.Strings.SearchValues;
 class FolkerKinzel.Strings.SearchValues<T>;